requirements.txt
luma.core==1.10.1
luma.oled==3.2.1
monotonic==1.5
Pillow==6.0.0
psutil==5.6.2
PyBluez==0.22
RPi.GPIO==0.6.5
smbus2==0.2.3
spidev==3.4
tcpbridge==1.1.1
pybluez[ble]==